Solution to Find the Median of an Array

+++ Find the Median of an Array@@676c02002dbc8c3e7a280b8c/code.c
@@ -19,8 +19,8 @@ int main(){
         printf("%d",arr[N/2]);
     }
     else{
-        double median = (arr[N/2] + arr[(N/2)-1])/2.0;
-        printf("%.1f",median);
+        int median = (arr[N/2] + arr[(N/2)-1])/2.0;
+        printf("%d",median);
     }
     return 0;
 }
